+++ examples/permissions/README.txt
@@ -1,4 +1,4 @@
 Set and query permissions
 ==========================
 
-This example script shows how to set and query permissions on files stored in remote backend.
+This example script shows how to set and query permissions on files stored in the remote backend.
