+++ examples/startup/data_transfer_runner.py
@@ -27,7 +27,7 @@
 Startup PyHPS Data Transfer
 ===========================
 
-Example script to start data transfer service client and query available storage.
+This example script starts a data transfer service client and queries available storage.
 
 Example usage:
 ``python examples/data_transfer_runner.py --debug``
@@ -62,7 +62,7 @@ def main(
     assert user_token is not None
 
     ##########################
-    # Create a Client instance
+    # Create a client instance
     # ========================
 
     client = Client()
