+++ examples/transfer_files/README.txt
@@ -1,5 +1,5 @@
 Transfer files using PyHPS Data Transfer
-===============================================
+========================================
 
-This example shows some of the main Data transfer client calls, like how to transfer files to remote backends,
-list files on a given path and how to download file from remote backends onto local storage.
+This example shows some of the main data transfer client calls, like how to transfer files to remote backends,
+list files on a given path, and download a file from remote backends onto local storage.
