+++ examples/transfer_files/transfer_files.py
@@ -23,16 +23,14 @@
 """
 .. _ref_transfer:
 
-==============================
-Basic script to transfer files
-==============================
+==============
+Transfer files
+==============
 
-Basic script to transfer files to remote backends and back using the data transfer service.
-'local-path' is the path to the files or directory to transfer. Supports wildcards.
+This basic script transfers files to remote backends and back using the data transfer service.
+The ``local-path`` argument is the path to the files or directory to transfer. Wildcards are supported.
 
-Example usage:
-
-    ``python examples/transfer_files.py --local_path=examples/basic/files/* --remote-path=hello --debug``
+Example usage: ``python examples/transfer_files.py --local_path=examples/basic/files/* --remote-path=hello --debug``
 
 """
 
@@ -49,8 +47,8 @@ import typer
 from typing_extensions import Annotated
 
 ###################################################
-# Necessary imports
-# =================
+# Perform necessary imports
+# =========================
 
 from ansys.hps.data_transfer.client import Client, DataTransferApi
 from ansys.hps.data_transfer.client.authenticate import authenticate
@@ -73,7 +71,7 @@ def transfer_files(api: DataTransferApi, local_path: str, remote_path: Optional[
     op = api.wait_for([op.id])
     log.debug(f"Operation {op[0].state}")
 
-    log.info("== Create remote directory ...")
+    log.info("== Creating remote directory ...")
     mkdir_op = api.mkdir([StoragePath(path=remote_path)])
     api.wait_for([mkdir_op.id])
     log.info(f"Directory {remote_path} created")
@@ -95,7 +93,7 @@ def transfer_files(api: DataTransferApi, local_path: str, remote_path: Optional[
     t1 = perf_counter()
     log.debug(f"Operation {op[0].state}")
 
-    log.info(f"== Query files and metadata in {remote_path} ...")
+    log.info(f"== Querying files and metadata in {remote_path} ...")
     op = api.list([StoragePath(path=remote_path)])
     op = api.wait_for([op.id])
     log.debug(f"Operation {op[0].state}")
@@ -169,7 +167,7 @@ def main(
 
     log.info("Connecting to the data transfer service client..")
     ##############################
-    # Create a ``client`` instance
+    # Create a client instance
     # ============================
     client = Client(clean=True)
