+++ src/ansys/hps/data_transfer/client/authenticate.py
@@ -49,7 +49,7 @@ def authenticate(
     verify: bool | str = True,
     **kwargs,
 ):
-    """Authenticates the user with a password or refresh token against the HPS authentication service.
+    """Authenticate the user with a password or refresh token against the HPS authentication service.
 
     If this method is successful, the response includes access and refresh tokens.
