+++ src/ansys/hps/data_transfer/client/client.py
@@ -71,7 +71,7 @@ def bin_in_use(bin_path):
 
 
 class MonitorState:
-    """Monitor and track state of worker binary."""
+    """Provides for monitoring and tracking the state of the worker binary."""
 
     def __init__(self):
         """Initialize the MonitorState class object."""
@@ -126,7 +126,7 @@ class MonitorState:
 
 
 class ClientBase:
-    """Provide the Python client to the HPS data transfer APIs.
+    """Provides the Python client to the HPS data transfer APIs.
 
     This class uses the provided credentials to create and store
     an authorized :class:`requests.Session` object.
@@ -134,23 +134,23 @@ class ClientBase:
     Parameters
     ----------
     bin_config: BinaryConfig
-        BinaryConfig object, Default is None
+        BinaryConfig object. The default is ``None``.
     download_dir: str
-        download dir path. Default is None
+        Download directory path. The default is ``None``.
     clean: bool
-        If True, cleans download_dir path. Default is False.
+        If True, cleans download_dir path. The default is ``False``..
     clean_dev: bool
-        If True, downloads binary from dev branch. Default is True.
+        Whether to download the binary from development branch. The default is ``True``.
     check_in_use: bool
-        If True, checks if binary is in use and skips downloading new binary. Default is True.
+        Whether to check if the binary is in use and skip downloading a new binary. The default is ``True``.
     timeout: float
-        Default is 60.0
+        The default is ``60.0``.
     retries: int
-        Default is 10
+        The default is ``1``.0
 
     Examples:
     --------
-    Create a client object and connect to HPS data transfer with a access_token.
+    Create a client object and connect to HPS data transfer with an access token.
 
     >>> from ansys.hps.data_transfer.client import Client
     >>> token = authenticate(username=username, password=password, verify=False, url=auth_url)
@@ -212,29 +212,29 @@ class ClientBase:
 
     @property
     def binary_config(self):
-        """Return binary config."""
+        """Binary configuration."""
         return self._bin_config
 
     @property
     def base_api_url(self):
-        """Return api url from config."""
+        """API URL from configuration."""
         return self._bin_config.url
 
     @property
     def session(self):
-        """Return a session object. Creates a new one if one is not already created."""
+        """Session object. Create a new one if one does not already exist."""
         if self._session is None:
             self._session = self._create_session(self.base_api_url, sync=not self.Meta.is_async)
         return self._session
 
     @property
     def is_started(self):
-        """Return True if binary is up and running."""
+        """Flag indicating if the binary is up and running."""
         return self.binary is not None and self.binary.is_started
 
     @property
     def timeout(self):
-        """Return the timeout for the session."""
+        """Timeout for the session."""
         return self._timeout
 
     @timeout.setter
@@ -244,7 +244,7 @@ class ClientBase:
 
     @property
     def retries(self):
-        """Return the number of retries."""
+        """Number of retries."""
         return self._retries
 
     @retries.setter
@@ -253,7 +253,7 @@ class ClientBase:
         self._retries = value
 
     def start(self):
-        """Start client session using binary config credentials."""
+        """Start client session using binary configuration credentials."""
         if self.binary is not None:
             return
 
@@ -276,7 +276,7 @@ class ClientBase:
         # self._session = self._create_session(self.base_api_url)
 
     def stop(self, wait=5.0):
-        """Stop client session."""
+        """Stop the client session."""
         if self.binary is None:
             return
         self._monitor_stop.set()
@@ -453,7 +453,7 @@ class ClientBase:
 
 
 class AsyncClient(ClientBase):
-    """An async interface to the Python client to the HPS data transfer APIs."""
+    """Provides an asynchronous interface to the Python client to the HPS data transfer APIs."""
 
     class Meta(ClientBase.Meta):
         """Meta class for AsyncClient class."""
@@ -535,7 +535,7 @@ class AsyncClient(ClientBase):
 
 
 class Client(ClientBase):
-    """Provide the Python client to the HPS data transfer APIs.
+    """Provides the Python client to the HPS data transfer APIs.
 
     This class uses the provided credentials to create and store
     an authorized :class:`requests.Session` object.
@@ -565,7 +565,7 @@ class Client(ClientBase):
         self._monitor_thread = None
 
     def start(self):
-        """Start client session using binary config credentials."""
+        """Start client session using binary configuration credentials."""
         super().start()
         atexit.register(self.stop)
         self._monitor_thread = threading.Thread(
@@ -574,7 +574,7 @@ class Client(ClientBase):
         self._monitor_thread.start()
 
     def stop(self, wait=5.0):
-        """Stop client session."""
+        """Stop the client session."""
         if self._session is not None:
             try:
                 self._session.post(self.base_api_url + "/shutdown")
@@ -583,7 +583,7 @@ class Client(ClientBase):
         super().stop(wait=wait)
 
     def wait(self, timeout: float = 60.0, sleep=0.5):
-        """Wait on worker binary to start."""
+        """Wait on the worker binary to start."""
         start = time.time()
         while time.time() - start < timeout:
             try:
