+++ src/ansys/hps/data_transfer/client/models/common.py
@@ -21,7 +21,7 @@
 # SOFTWARE.
 
 # generated by datamodel-codegen
-"""Module providing UserInfo schema."""
+"""Module providing the ``UserInfo`` schema."""
 
 from __future__ import annotations
