+++ src/ansys/hps/data_transfer/client/models/metadata.py
@@ -21,7 +21,7 @@
 # SOFTWARE.
 
 # generated by datamodel-codegen
-"""Module providing DataAssignment, DataAssignments and SrcDst schema."""
+"""Module providing ``DataAssignment``, ``DataAssignments``, and ``SrcDst`` schemas."""
 
 from __future__ import annotations
