+++ src/ansys/hps/data_transfer/client/models/msg.py
@@ -21,7 +21,7 @@
 # SOFTWARE.
 
 # generated by datamodel-codegen
-"""Module providing BuildInfo schema and a other data transfer operation Response schemas."""
+"""Module providing ``BuildInfo`` schema and other data transfer operation ``Response`` schemas."""
 
 from __future__ import annotations
