+++ src/ansys/hps/data_transfer/client/models/ops.py
@@ -21,7 +21,7 @@
 # SOFTWARE.
 
 # generated by datamodel-codegen
-"""Module providing OperationState and Operation schemas."""
+"""Module providing ``OperationState`` and ``Operation`` schemas."""
 
 from __future__ import annotations
