+++ src/ansys/hps/data_transfer/client/models/permissions.py
@@ -21,7 +21,7 @@
 # SOFTWARE.
 
 # generated by datamodel-codegen
-"""Module providing ResourceType, RoleType and SubjectType schema.
+"""Module providing ``ResourceType``, ``RoleType``, and ``SubjectType`` schemas.
 
 Permission functionality related RoleAssignment and  RoleQuery schema.
 """
