+++ src/ansys/hps/data_transfer/client/models/storage.py
@@ -21,7 +21,7 @@
 # SOFTWARE.
 
 # generated by datamodel-codegen
-"""Module providing PathType schema."""
+"""Module providing the ``PathType`` schema."""
 
 from __future__ import annotations
